+++ flutter_application_1/lib/Data/destinations.dart
@@ -1,0 +1,73 @@
+import 'package:flutter_application_1/Models/destination.dart';
+
+final List<Destination> destinations = [
+  Destination(
+    city: 'Kandy',
+    province: 'Central Province',
+    activities: 10,
+    description: 'Beautiful city with rich culture and history.',
+    imagePath: "assets/kandy.jpg",
+  ),
+  Destination(
+    city: 'Galle',
+    province: 'Southern Province',
+    activities: 8,
+    description: 'Historical city with a stunning fort.',
+    imagePath: "assets/galle2.jpg",
+  ),
+  Destination(
+    city: 'Colombo',
+    province: 'Western Province',
+    activities: 12,
+    description: 'Capital city with vibrant nightlife and shopping.',
+    imagePath: "assets/cmb.jpg",
+  ),
+  Destination(
+    city: 'Sigiriya',
+    province: 'Central Province',
+    activities: 6,
+    description: 'Home to the ancient rock fortress of Sigiriya.',
+    imagePath: "assets/sigiriya2.jpg",
+  ),
+  Destination(
+      city: 'Nuwara Eliya',
+      province: 'Central Province',
+      activities: 7,
+      description:
+          'Scenic city known for its tea plantations and cool climate.',
+      imagePath: "assets/ne.jpg"),
+  Destination(
+      city: 'Anuradhapura',
+      province: 'North Central Province',
+      activities: 10,
+      description:
+          'Anuradhapura, a Ceylonese political and religious capital that flourished for 1,300 years, was abandoned after an invasion in 993.',
+      imagePath: "assets/02.jpg"),
+  Destination(
+      city: 'Polonnaruwa',
+      province: 'North Central Province',
+      activities: 9,
+      description:
+          'The second oldest of all Sri Lankas kingdoms, Polonnaruwa was first established as a military post by the Sinhalese kingdom.',
+      imagePath: "assets/pol.jpg"),
+  Destination(
+      city: 'Matara',
+      province: 'Southern Province',
+      activities: 5,
+      description:
+          'Matara is a major city in Sri Lanka, on the southern coast of Southern Province.',
+      imagePath: "assets/mat.jpg"),
+  Destination(
+      city: 'Jaffna',
+      province: 'North Province',
+      activities: 10,
+      description: 'Jaffna is a city on the northern tip of Sri Lanka.',
+      imagePath: "assets/jaffna.jpg"),
+  Destination(
+      city: 'Trincomalee',
+      province: 'Estern Province',
+      activities: 7,
+      description:
+          'Trincomalee is a port city on the northeast coast of Sri Lanka.',
+      imagePath: "assets/tri.jpg"),
+];
