+++ flutter_application_1/lib/Data/unbordingContents.dart
@@ -1,11 +1,5 @@
-class UnbordingContent {
-  String image;
-  String title;
-  String description;
-
-  UnbordingContent(
-      {required this.image, required this.title, required this.description});
-}
+// ignore: file_names
+import 'package:flutter_application_1/Models/unbordingContent.dart';
 
 List<UnbordingContent> contents = [
   UnbordingContent(
