+++ flutter_application_1/lib/Models/destination.dart
@@ -1,0 +1,15 @@
+class Destination {
+  final String city;
+  final int? activities;
+  final String description;
+  final String province;
+  final String imagePath;
+
+  Destination({
+    required this.city,
+    this.activities,
+    this.description = '',
+    required this.province,
+    required this.imagePath,
+  });
+}
