+++ flutter_application_1/lib/Models/hotel.dart
@@ -1,0 +1,13 @@
+class Hotel {
+  final String name;
+  final String address;
+  final double price;
+  final String imagePath;
+
+  Hotel({
+    required this.name,
+    required this.address,
+    required this.price,
+    required this.imagePath,
+  });
+}
