+++ flutter_application_1/lib/Models/unbordingContent.dart
@@ -1,0 +1,12 @@
+// ignore: file_names
+class UnbordingContent {
+  String image;
+  String title;
+  String description;
+
+  UnbordingContent({
+    required this.image,
+    required this.title,
+    required this.description,
+  });
+}
