+++ flutter_application_1/lib/screens/destination_screen.dart
@@ -1,14 +1,19 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_application_1/Models/destination.dart';
 import 'package:flutter_application_1/screens/activity_screen.dart';
-import 'package:flutter_application_1/widgets/activity.dart';
-import 'package:flutter_application_1/widgets/destination_carousel.dart';
+import 'package:flutter_application_1/Data/activities.dart';
+import 'package:flutter_application_1/Models/activity.dart';
 
 class DestinationScreen extends StatefulWidget {
   final Destination destination;
 
-  const DestinationScreen({required this.destination});
+  const DestinationScreen({
+    super.key,
+    required this.destination,
+  });
 
   @override
+  // ignore: library_private_types_in_public_api
   _DestinationScreenState createState() => _DestinationScreenState();
 }
 
@@ -33,7 +38,7 @@ class _DestinationScreenState extends State<DestinationScreen> {
                 height: MediaQuery.of(context).size.width,
                 decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(30),
-                  boxShadow: [
+                  boxShadow: const [
                     BoxShadow(
                       color: Colors.black26,
                       offset: Offset(0.0, 2.0),
@@ -53,13 +58,13 @@ class _DestinationScreenState extends State<DestinationScreen> {
                 ),
               ),
               Padding(
-                padding: EdgeInsets.symmetric(horizontal: 10, vertical: 40),
+                padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 40),
                 child: Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: <Widget>[
                     IconButton(
                       onPressed: () => Navigator.pop(context),
-                      icon: Icon(Icons.arrow_back_rounded),
+                      icon: const Icon(Icons.arrow_back_rounded),
                       iconSize: 30,
                       color: Colors.white,
                     ),
@@ -67,13 +72,13 @@ class _DestinationScreenState extends State<DestinationScreen> {
                       children: <Widget>[
                         IconButton(
                           onPressed: () => Navigator.pop(context),
-                          icon: Icon(Icons.search_rounded),
+                          icon: const Icon(Icons.search_rounded),
                           iconSize: 30,
                           color: Colors.white,
                         ),
                         IconButton(
                           onPressed: () => Navigator.pop(context),
-                          icon: Icon(Icons.sort_rounded),
+                          icon: const Icon(Icons.sort_rounded),
                           iconSize: 30,
                           color: Colors.white,
                         ),
@@ -138,7 +143,7 @@ class _DestinationScreenState extends State<DestinationScreen> {
           ),
           Expanded(
             child: ListView.builder(
-              padding: EdgeInsets.only(top: 10.0, bottom: 15.0),
+              padding: const EdgeInsets.only(top: 10.0, bottom: 15.0),
               itemCount: 10,
               itemBuilder: (BuildContext context, int index) {
                 Activity activity = activities[index];
@@ -154,7 +159,7 @@ class _DestinationScreenState extends State<DestinationScreen> {
                   child: Stack(
                     children: [
                       Container(
-                        margin: EdgeInsets.fromLTRB(40, 5, 20, 5),
+                        margin: const EdgeInsets.fromLTRB(40, 5, 20, 5),
                         height: 170,
                         width: double.infinity,
                         decoration: BoxDecoration(
@@ -180,7 +185,7 @@ class _DestinationScreenState extends State<DestinationScreen> {
                                     MainAxisAlignment.spaceBetween,
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: [
-                                  Container(
+                                  SizedBox(
                                     width: 120.0,
                                     child: Text(
                                       activity.name,
@@ -225,12 +230,12 @@ class _DestinationScreenState extends State<DestinationScreen> {
                                 ),
                               ),
                               _buildRatingStars(activity.rating),
-                              SizedBox(height: 5),
+                              const SizedBox(height: 5),
                               Row(
                                 children: [
-                                  if (activity.startTimes.length > 0) ...[
+                                  if (activity.startTimes.isNotEmpty) ...[
                                     Container(
-                                      padding: EdgeInsets.all(1.0),
+                                      padding: const EdgeInsets.all(1.0),
                                       width: 60,
                                       decoration: BoxDecoration(
                                         color: const Color.fromARGB(
@@ -248,11 +253,11 @@ class _DestinationScreenState extends State<DestinationScreen> {
                                         ),
                                       ),
                                     ),
-                                    SizedBox(width: 5.0),
+                                    const SizedBox(width: 5.0),
                                   ],
                                   if (activity.startTimes.length > 1) ...[
                                     Container(
-                                      padding: EdgeInsets.all(1.0),
+                                      padding: const EdgeInsets.all(1.0),
                                       width: 60,
                                       decoration: BoxDecoration(
                                         color: const Color.fromARGB(
