+++ flutter_application_1/lib/widgets/hotels_carousel.dart
@@ -1,86 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_application_1/Data/hotels.dart';
+import 'package:flutter_application_1/Models/hotel.dart';
 // ignore: unused_import
 import 'package:flutter_application_1/home_screen.dart';
 import 'package:flutter_application_1/screens/hotel_screen.dart';
 
-class Hotel {
-  final String name;
-  final String address;
-  final double price;
-  final String imagePath;
 
-  Hotel({
-    required this.name,
-    required this.address,
-    required this.price,
-    required this.imagePath,
-  });
-}
 
-final List<Hotel> hotels = [
-  Hotel(
-    name: 'Wild Coast Tented Lodge',
-    address: 'Wild Coast Tented Lodge Palatupana, Yala.',
-    price: 1245,
-    imagePath: "assets/h11.jpg",
-  ),
-  Hotel(
-    name: 'Le Grand Galle',
-    address: 'Hilltop Buiding, 30 Park Rd, Galle.',
-    price: 223,
-    imagePath: "assets/h2.jpg",
-  ),
-  Hotel(
-    name: 'Shangri-La Hambantota',
-    address: 'Sittrakala Watta, Chithragala, Ambalantota.',
-    price: 230,
-    imagePath: "assets/h3.jpg",
-  ),
-  Hotel(
-    name: 'Jetwing Lake',
-    address: 'Wijaya Kumaratunga Mawatha Yapagama, Dambulla',
-    price: 125,
-    imagePath: "assets/h4.jpg",
-  ),
-  Hotel(
-    name: 'The Golden Ridge Hotel',
-    address: '395, Road, Nuwaraeliya 22200',
-    price: 145,
-    imagePath: "assets/h5.jpg",
-  ),
-  Hotel(
-    name: 'Jetwing Blue',
-    address: '6RRR+MPF, Porutota Rd, Negombo 11500',
-    price: 140,
-    imagePath: "assets/h6.jpg",
-  ),
-  Hotel(
-    name: 'Jetwing Surf',
-    address: 'P20, Kottukal Road, Hidayapuram, Pottuvil 32500',
-    price: 148,
-    imagePath: "assets/h7.jpg",
-  ),
-  Hotel(
-    name: 'Taj Bentota Resort & Spa',
-    address: 'National Holiday Resort, Galle, Bentota 80500',
-    price: 218,
-    imagePath: "assets/h8.jpg",
-  ),
-  Hotel(
-    name: 'Heritance Ahungalla',
-    address: 'Galle Rd, Ahungalla',
-    price: 175,
-    imagePath: "assets/h9.jpg",
-  ),
-  Hotel(
-    name: 'Wattura Resort and Spa',
-    address: '1271 Kammala South Road, Waikkal 61110',
-    price: 265,
-    imagePath: "assets/h10.jpg",
-  ),
-];
+
 
 class HotelCarousel extends StatelessWidget {
+  const HotelCarousel({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Column(
